Mark sms sent status failed when any exceptions occur
